Added Turrets, Subcrafts that are part of a larger craft but can be rotated independantly using the new "subcraft rotate" signs
Fixed the bug where redstone drops off dispensers. This time for realz!
...maybe

Made it so it tells you which block is causing an obstruction

Made it so it tells the console when a craft with no pilot is released

Rolled some of the performance improvements already in the main updater
into the compatibilitymode updater