Update pom.xml
I added the Spigot (http://www.spigotmc.org) repository and dependencies in the Maven file, which is where future Minecraft 1.8 support is going to be coming from. By building against Spigot, you are using the API that most of the Minecraft community is headed towards right now. I personally think that if this project is going to continue, this is definitely a needed change!

Additionally, I also changed the version of Plugin Metrics from R6 to R8-SNAPSHOT, the standard version for Spigot 1.8.